houseparty/requirements.txt
jira
todoist-python
